Refactoring SetupManager related to #932
The setup manager is quite complex. We are currently solving a bug related to the fact
that changes in ingress also change the GSLB and can cause cycles #932. I am doing the
necessary refactoring (code is better structured, but the functionality remains the same),
to make it easier to make the necessary changes..

 - split SetupWithManager and Reconciliation into two files
 - encapsulate createGslbFromIngress
 - simplify ingressHandler
 - extracting parse strategy func
