+++ controllers/gslb_controller_reconciliation.go
@@ -21,7 +21,6 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	"context"
 	"fmt"
-	"strconv"
 
 	"github.com/k8gb-io/k8gb/controllers/providers/metrics"
 
@@ -32,19 +31,10 @@ import (
 	"github.com/k8gb-io/k8gb/controllers/providers/dns"
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/trace"
-	corev1 "k8s.io/api/core/v1"
-	netv1 "k8s.io/api/networking/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	types "k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	"sigs.k8s.io/controller-runtime/pkg/handler"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	"sigs.k8s.io/controller-runtime/pkg/source"
-	externaldns "sigs.k8s.io/external-dns/endpoint"
 )
 
 // GslbReconciler reconciles a Gslb object
@@ -197,159 +187,4 @@ func (r *GslbReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.
 	// TODO: potentially enhance with smarter reaction to external Event
 	m.IncrementReconciliation(gslb)
 	return result.Requeue()
-}
-
-// SetupWithManager configures controller manager
-func (r *GslbReconciler) SetupWithManager(mgr ctrl.Manager) error {
-	// Figure out Gslb resource name to Reconcile when non controlled Name is updated
-
-	endpointMapHandler := handler.EnqueueRequestsFromMapFunc(
-		func(a client.Object) []reconcile.Request {
-			gslbList := &k8gbv1beta1.GslbList{}
-			opts := []client.ListOption{
-				client.InNamespace(a.GetNamespace()),
-			}
-			c := mgr.GetClient()
-			err := c.List(context.TODO(), gslbList, opts...)
-			if err != nil {
-				log.Info().Msg("Can't fetch gslb objects")
-				return nil
-			}
-			gslbName := ""
-			for _, gslb := range gslbList.Items {
-				for _, rule := range gslb.Spec.Ingress.Rules {
-					for _, path := range rule.HTTP.Paths {
-						if path.Backend.Service != nil && path.Backend.Service.Name == a.GetName() {
-							gslbName = gslb.Name
-						}
-					}
-				}
-			}
-			if len(gslbName) > 0 {
-				return []reconcile.Request{
-					{NamespacedName: types.NamespacedName{
-						Name:      gslbName,
-						Namespace: a.GetNamespace(),
-					}},
-				}
-			}
-			return nil
-		})
-
-	createGslbFromIngress := func(annotationKey string, annotationValue string, a client.Object, strategy string) {
-		log.Info().
-			Str("annotation", fmt.Sprintf("(%s:%s)", annotationKey, annotationValue)).
-			Str("ingress", a.GetName()).
-			Msg("Detected strategy annotation on ingress")
-		c := mgr.GetClient()
-		ingressToReuse := &netv1.Ingress{}
-		err := c.Get(context.Background(), client.ObjectKey{
-			Namespace: a.GetNamespace(),
-			Name:      a.GetName(),
-		}, ingressToReuse)
-		if err != nil {
-			log.Info().
-				Str("ingress", a.GetName()).
-				Msg("Ingress does not exist anymore. Skipping Glsb creation...")
-			return
-		}
-		gslbExist := &k8gbv1beta1.Gslb{}
-		err = c.Get(context.Background(), client.ObjectKey{
-			Namespace: a.GetNamespace(),
-			Name:      a.GetName(),
-		}, gslbExist)
-		if err == nil {
-			log.Info().
-				Str("gslb", gslbExist.Name).
-				Msg("Gslb already exists. Skipping Gslb creation...")
-			return
-		}
-		gslb := &k8gbv1beta1.Gslb{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace:   a.GetNamespace(),
-				Name:        a.GetName(),
-				Annotations: a.GetAnnotations(),
-			},
-			Spec: k8gbv1beta1.GslbSpec{
-				Ingress: k8gbv1beta1.FromV1IngressSpec(ingressToReuse.Spec),
-				Strategy: k8gbv1beta1.Strategy{
-					Type: strategy,
-				},
-			},
-		}
-
-		annotationToInt := func(k string, v string) int {
-			intValue, err := strconv.Atoi(v)
-			if err != nil {
-				log.Err(err).
-					Str("annotationKey", k).
-					Str("annotationValue", v).
-					Msg("Can't parse annotation value to int")
-			}
-			return intValue
-		}
-
-		for annotationKey, annotationValue := range a.GetAnnotations() {
-			switch annotationKey {
-			case dnsTTLSecondsAnnotation:
-				gslb.Spec.Strategy.DNSTtlSeconds = annotationToInt(annotationKey, annotationValue)
-			case splitBrainThresholdSecondsAnnotation:
-				gslb.Spec.Strategy.SplitBrainThresholdSeconds = annotationToInt(annotationKey, annotationValue)
-			}
-		}
-
-		if strategy == depresolver.FailoverStrategy {
-			for annotationKey, annotationValue := range a.GetAnnotations() {
-				if annotationKey == primaryGeoTagAnnotation {
-					gslb.Spec.Strategy.PrimaryGeoTag = annotationValue
-				}
-			}
-			if gslb.Spec.Strategy.PrimaryGeoTag == "" {
-				log.Info().
-					Str("annotation", primaryGeoTagAnnotation).
-					Str("gslb", gslb.Name).
-					Msg("Annotation is missing, skipping Gslb creation...")
-				return
-			}
-		}
-
-		err = controllerutil.SetControllerReference(ingressToReuse, gslb, r.Scheme)
-		if err != nil {
-			log.Err(err).
-				Str("ingress", ingressToReuse.Name).
-				Str("gslb", gslb.Name).
-				Msg("Cannot set the Ingress as the owner of the Gslb")
-		}
-
-		log.Info().
-			Str("gslb", gslb.Name).
-			Msg("Creating new Gslb out of Ingress annotation")
-		err = c.Create(context.Background(), gslb)
-		if err != nil {
-			log.Err(err).Msg("Glsb creation failed")
-		}
-	}
-
-	ingressMapHandler := handler.EnqueueRequestsFromMapFunc(
-		func(a client.Object) []reconcile.Request {
-			for annotationKey, annotationValue := range a.GetAnnotations() {
-				if annotationKey == strategyAnnotation {
-					switch annotationValue {
-					case depresolver.RoundRobinStrategy:
-						createGslbFromIngress(annotationKey, annotationValue, a, depresolver.RoundRobinStrategy)
-					case depresolver.FailoverStrategy:
-						createGslbFromIngress(annotationKey, annotationValue, a, depresolver.FailoverStrategy)
-					}
-				}
-			}
-			return nil
-		})
-
-	return ctrl.NewControllerManagedBy(mgr).
-		For(&k8gbv1beta1.Gslb{}).
-		Owns(&netv1.Ingress{}).
-		Owns(&externaldns.DNSEndpoint{}).
-		Watches(&source.Kind{Type: &corev1.Endpoints{}}, endpointMapHandler).
-		Watches(&source.Kind{Type: &netv1.Ingress{}}, ingressMapHandler).
-		Complete(r)
 }
